source/requirements.txt
factory_boy==2.11.1
flask==1.0.2
marshmallow==2.16.3
mysql-connector-python==8.0.5
pytest==3.10.0
SQLAlchemy==1.2.13
